#3580 cart trash button disabled when no selection

+++ src/sets/components/cart-documents-table/component.ts
@@ -119,6 +119,9 @@ class CartDocumentsTableController implements ng.IComponentController {
   }
 
   public clearSet(): void {
+    if (!this.hasSelections()) {
+      return;
+    }
     const sels = this.getSelectedDocumentIds();
     if (sels && sels.length > 0) {
       this.SetService.removeDocumentFromCart(this.type, sels)
